Clean up some methods
